+++ app/controllers/application_controller.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
 class ApplicationController < ActionController::Base
-  before_action :configure_permitted_parameters, if: :devise_controller?
-  before_action :authenticate_user!
+  include UserAuthorization
+  include DeviseOverrides
 
   def paginate_and_filter_jobs(jobs)
     @status = params[:status]
@@ -10,16 +10,5 @@ class ApplicationController < ActionController::Base
     jobs = jobs.where(status: @status) if @status
 
     jobs
-  end
-
-  def after_invite_path_for(_inviter, _invitee = nil)
-    users_path
-  end
-
-  protected
-
-  def configure_permitted_parameters
-    devise_parameter_sanitizer.permit(:invite, keys: %i[username email])
-    devise_parameter_sanitizer.permit(:account_update, keys: %i[username])
   end
 end
