+++ app/controllers/users_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class UsersController < ApplicationController
+  before_action :authenticate_admin!
+
   def index
     @users = User.order(username: :desc).page(params[:page])
   end
