+++ app/models/user.rb
@@ -4,4 +4,6 @@ class User < ApplicationRecord
   # Include default devise modules. Others available are:
   # :confirmable, :lockable, :timeoutable, :trackable and :omniauthable
   devise :database_authenticatable, :recoverable, :rememberable, :validatable, :invitable
+
+  enum :role, { harvester: 0, admin: 1 }
 end
