+++ app/views/devise/passwords/new.html.erb
@@ -22,7 +22,7 @@
             <%= f.submit "Send me reset password instructions", class: "btn btn-primary" %>
           </div>
 
-        <%= render "devise/shared/links" %>
+          <%= render "devise/shared/links" %>
         </div>
       </div>
     <% end %>
