+++ app/views/shared/_nav.html.erb
@@ -43,7 +43,9 @@
             </a>
             <ul class="dropdown-menu">
               <li><%= link_to 'Edit profile', edit_user_registration_path, class: "dropdown-item" %></li>
-              <li><%= link_to 'Users', users_path, class: "dropdown-item" %></li>
+              <% if current_user.admin? %>
+                <li><%= link_to 'Users', users_path, class: "dropdown-item" %></li>
+              <% end %>
               <li><hr class="dropdown-divider"></li>
               <li><%= button_to 'Logout', destroy_user_session_path, method: :delete, class: "dropdown-item" %></li>
             </ul>
