+++ db/migrate/20230605235157_add_role_to_users.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddRoleToUsers < ActiveRecord::Migration[7.0]
+  def change
+    add_column :users, :role, :integer, null: false, default: 0
+  end
+end
